+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/service/credential/DeferredCredentialService.java
@@ -160,8 +160,6 @@ public class DeferredCredentialService {
 
         credentialOfferRepository.save(credentialOffer);
         credentialManagementRepository.save(mgmt);
-
-        eventProducerService.produceOfferStateChangeEvent(mgmt.getId(), credentialOffer.getId(), credentialOffer.getCredentialStatus());
     }
 
     /**