+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/service/webhook/EventProducerService.java
@@ -35,21 +35,30 @@ public class EventProducerService {
         applicationEventPublisher.publishEvent(errorEvent);
     }
 
-    public void produceStateChangeEvent(UUID credentialOfferId, CredentialStatusManagementType state) {
-        var stateChangeEvent = new StateChangeEvent(
-                credentialOfferId,
+    /**
+     * Only used for Credential Management State Changes from CredentialStateMachineAction
+     * @param credentialManagementId
+     * @param state
+     */
+    public void produceManagementStateChangeEvent(UUID credentialManagementId, CredentialStatusManagementType state) {
+        var managementStateChangeEvent = new ManagementStateChangeEvent(
+                credentialManagementId,
                 state
         );
-        applicationEventPublisher.publishEvent(stateChangeEvent);
+        applicationEventPublisher.publishEvent(managementStateChangeEvent);
     }
 
-    public void produceOfferStateChangeEvent(UUID credentialManagementId, UUID credentialOfferId, CredentialOfferStatusType state) {
-        var stateChangeEvent = new OfferStateChangeEvent(
-                credentialManagementId,
+    /**
+     * Only used for Credential Offer State Changes from CredentialStateMachineAction
+     * @param credentialOfferId
+     * @param state
+     */
+    public void produceOfferStateChangeEvent(UUID credentialOfferId, CredentialOfferStatusType state) {
+        var offerStateChangeEvent = new OfferStateChangeEvent(
                 credentialOfferId,
                 state
         );
-        applicationEventPublisher.publishEvent(stateChangeEvent);
+        applicationEventPublisher.publishEvent(offerStateChangeEvent);
     }
 
     public void produceDeferredEvent(CredentialOffer credentialOffer, ClientAgentInfo clientInfo) {