+++ issuer-service/src/test/java/ch/admin/bj/swiyu/issuer/service/credential/CredentialEnvelopeServiceTest.java
@@ -101,7 +101,6 @@ class CredentialEnvelopeServiceTest {
         verify(credentialStateMachine).sendEventAndUpdateStatus(mgmt, CredentialStateMachineConfig.CredentialManagementEvent.ISSUE);
         verify(credentialOfferRepository).save(offer);
         verify(credentialManagementRepository).save(mgmt);
-        verify(eventProducerService).produceOfferStateChangeEvent(mgmt.getId(), offer.getId(), offer.getCredentialStatus());
     }
 
     @Test
@@ -136,7 +135,6 @@ class CredentialEnvelopeServiceTest {
         verify(credentialStateMachine).sendEventAndUpdateStatus(mgmt, CredentialStateMachineConfig.CredentialManagementEvent.ISSUE);
         verify(credentialOfferRepository).save(offer);
         verify(credentialManagementRepository).save(mgmt);
-        verify(eventProducerService).produceOfferStateChangeEvent(mgmt.getId(), offer.getId(), offer.getCredentialStatus());
     }